feat: add TypeScript posts page
