+++ client-react/package.json
@@ -6,5 +6,13 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "bits-ui": "^1.0.0"
+  },
+  "devDependencies": {
+    "@types/react": "^18.2.0",
+    "@types/react-dom": "^18.2.0",
+    "typescript": "^5.4.0"
+  },
+  "scripts": {
+    "build": "tsc --noEmit"
   }
 }
